requirements.txt
streamlit
pytesseract
pdf2image
pillow
pandas
